Add Slot to Replace VDateInput Action slot
I need to update the VDateInput so that I can manipulate the actions on the VDateInput. Turns out its the VConfirmEdit component. 

So what we do is we check if slots.actions is defined and if it is we pass it to the actions slot of VDatePicker, along with the necessary slot props (actions and model) and if its not, we use the default actions provided by VConfirmEdit. now i can use VDateInput and slotcustom actions slot to replace the default actions in VDatePicker like so

<template>
  <VDateInput v-model="selectedDate">
    <template #actions="{ actions, model }">
      <VBtn text @click="setTodaysDate()">Today</VBtn>
    </template>
  </VDateInput>
</template>